Incorporate github.com/goccy/go-json as an optional module
I still feel that github.com/goccy/go-json is a little bit unstable,
but the speed benefits are just too great to ignore.

Now encoding/json is the default library being used, but if
serialization speed is a problem, you can try enabling the fater
library by build with the "goccy" tag

  go build -tags goccy ....

In some cases you get up to 70% performance increase!
